requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
matplotlib>=3.5.0
tkinterdnd2>=0.3
